+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.8] - 2026-04-20
+
+### Docs
+- Update README.md
+- Update examples/md/04-v3-state-reconciliation/migration.md
+
+### Other
+- Update redeploy/cli.py
+- Update redeploy/tests/test_version_cli.py
+
 ## [0.2.7] - 2026-04-20
 
 ### Docs